WIP: Allow setting and reseting bool values via standard arguments
This commit enables the developer to allow setting boolean values to
`false`. There already was a piece of code which indirectly allowed
this, but default could never be true for bools due to #159.

With this patch, a new setting `disable-type` is introduced on boolean
options. If unset, the bool variables act as before. However, if set,
the user can set (and reset) the bool varaible to either `true` or
`false`.

The setting has three possible values: `value`. `no` and
`enable-disable`.

When the setting is enabled, bool values will behave as follows:

| *`disable-value`* | Syntax                    |
| `value`           | `--bool[=true|false]`     |
| `no`              | `--[no-]bool`             |
| `enable-disable`  | `--[enable|disable]-bool` |

This allows for use cases such as:

```sh
alias program='program --debug`
program --no-debug
```

or

```sh
./program --enable-raytracing --disable-antialiasing
```